refactor(survicate): componentize survicate widget
